+++ examples/server_profile_template.go
@@ -11,7 +11,7 @@ func main() {
 	var (
 		clientOV                     *ov.OVClient
 		server_profile_template_name = "Test SPT"
-		enclosure_group_name         = "RenamedEnclosureGroup"
+		enclosure_group_name         = "enclosureGp"
 		server_hardware_type_name    = "SY 480 Gen9 1"
 	)
 	ovc := clientOV.NewOVClient(
@@ -20,7 +20,7 @@ func main() {
 		os.Getenv("ONEVIEW_OV_DOMAIN"),
 		os.Getenv("ONEVIEW_OV_ENDPOINT"),
 		false,
-		1800,
+		2000,
 		"*")
 
 	server_hardware_type, err := ovc.GetServerHardwareTypeByName(server_hardware_type_name)
@@ -31,7 +31,10 @@ func main() {
 		ManageConnections: true,
 	}
 	initialScopeUris := new([]utils.Nstring)
-	*initialScopeUris = append(*initialScopeUris, utils.NewNstring("/rest/scopes/7fe26585-b7a1-497e-992e-90908f70dfaf"))
+	scp, scperr := ovc.GetScopeByName("test")
+	if scperr != nil {
+		*initialScopeUris = append(*initialScopeUris, scp.URI)
+	}
 
 	server_profile_template := ov.ServerProfile{
 		Type:                  "ServerProfileTemplateV8",